+++ src/main/kotlin/graphql/dataTransferObjects/CannotLeaveChat.kt
@@ -2,6 +2,6 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
-object CannotLeaveChat : LeaveGroupChatResult {
+object CannotLeaveChat : LeaveGroupChatResult, RemoveGroupChatUsersResult {
     fun getPlaceholder() = Placeholder
 }