+++ src/main/kotlin/graphql/dataTransferObjects/MustBeAdmin.kt
@@ -2,6 +2,6 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
-object MustBeAdmin {
+object MustBeAdmin : RemoveGroupChatUsersResult {
     fun getPlaceholder() = Placeholder
 }